Updated version and adding hardening for possible infinite loop
If scheduled recording LastUpdate is in the future (say system time was
specified accidentally instead of UTC time or some other similar mistake
was made), the service could attempt to sync infinitely so this forces
an update to the LastUpdate time to avoid this case. Also rev'ed the
version in preparation for release and tweaked a few strings.