Initial support for renaming an existing session and rescheduling a moved session
Now can rename a session if session name changed and last update time
was increased. Also can "move" an existing session if the RR changes and
last update time was increased. Basically checks if the RR changed then
deletes the old session and creates a new session. Also added support
for recording to the RR's default folder if an invalid folder is
provided.